refactor: Chat
